requirements.txt
requests==2.31.0
pandas==2.1.0
psycopg2-binary==2.9.9
boto3==1.34.0
python-dotenv==1.0.0